+++ apps/storybook/.storybook/preview.tsx
@@ -1,9 +1,13 @@
 import React from "react";
 import { withThemeByClassName } from "@storybook/addon-themes";
 import type { Preview, ReactRenderer } from "@storybook/react";
+import { initialize, mswLoader } from "msw-storybook-addon";
 import { Toaster } from "sonner";
 
 import "../src/app/globals.css";
+
+// Initialize MSW
+initialize({ onUnhandledRequest: "bypass" });
 
 const preview: Preview = {
   parameters: {
@@ -16,6 +20,7 @@ const preview: Preview = {
     },
     docs: { toc: true },
   },
+  loaders: [mswLoader],
   decorators: [
     (Story) => (
       <>