+++ apps/storybook/next.config.js
@@ -3,7 +3,7 @@ const config = {
   reactStrictMode: true,
 
   /** Enables hot reloading for local packages without a build step */
-  transpilePackages: ["@swy/ui"],
+  transpilePackages: ["@swy/ui", "@swy/liveblocks"],
 
   /** We already do linting and typechecking as separate tasks in CI */
   eslint: { ignoreDuringBuilds: true },