+++ apps/worxpace/src/app/(platform)/(tools)/layout.tsx
@@ -1,41 +1,42 @@
 "use client";
 
-import { type PropsWithChildren } from "react";
-import { useParams, useRouter } from "next/navigation";
-import { useAuth } from "@clerk/nextjs";
+import { useCallback, useMemo, type PropsWithChildren } from "react";
+import { useParams } from "next/navigation";
 
 import { TreeProvider } from "@swy/ui/custom";
 
 import { GROUPS } from "~/constants/site";
-import { usePlatform } from "~/hooks";
+import { useDocuments, usePlatform } from "~/hooks";
+import { toTreeItem } from "~/lib";
 import WorkspaceLayout from "./_components/workspace-layout";
 
 const ToolsLayout = ({ children }: PropsWithChildren) => {
-  const { isSignedIn } = useAuth();
-  const { toToolsPage } = usePlatform();
+  const { toToolsPage, workspaceId } = usePlatform();
   /** Routing */
-  const router = useRouter();
   const params = useParams<{
     documentId?: string;
     boardId?: string;
     whiteboardId?: string;
     workflowId?: string;
   }>();
-  const pageId =
-    params.documentId ??
-    params.boardId ??
-    params.whiteboardId ??
-    params.workflowId ??
-    null;
-  const isItemActive = (id: string) => {
-    return pageId === id;
-  };
+  const pageId = useMemo(
+    () =>
+      params.documentId ??
+      params.boardId ??
+      params.whiteboardId ??
+      params.workflowId ??
+      null,
+    [params],
+  );
+  const isItemActive = useCallback((id: string) => pageId === id, [pageId]);
+  const { documents } = useDocuments({ workspaceId });
+  const pages = useMemo(() => documents?.map(toTreeItem) ?? [], [documents]);
 
-  if (!isSignedIn) router.push("/select-role");
   return (
     <TreeProvider
       className="flex h-screen bg-main"
       groups={GROUPS}
+      initialItems={pages}
       onClickItem={toToolsPage}
       isItemActive={isItemActive}
     >